Extract dcos interface wiht DCOSTools and Finder
