+++ api/pull.go
@@ -2,261 +2,16 @@ package api
 
 import (
 	"encoding/json"
-	"errors"
 	"fmt"
-	"io/ioutil"
-	"net"
 	"net/http"
-	"path/filepath"
-	"strconv"
 	"sync"
 	"time"
 
 	"github.com/dcos/dcos-diagnostics/config"
 	"github.com/dcos/dcos-diagnostics/dcos"
+	"github.com/dcos/dcos-diagnostics/util"
 	"github.com/sirupsen/logrus"
 )
-
-// with the nodeFinder interface we can chain finding methods
-type nodeFinder interface {
-	find() ([]dcos.Node, error)
-}
-
-// find masters via dns. Used to find master nodes from agents.
-type findMastersInExhibitor struct {
-	url  string
-	next nodeFinder
-
-	// getFn takes url and timeout and returns a read body, HTTP status code and error.
-	getFn func(string, time.Duration) ([]byte, int, error)
-}
-
-func (f *findMastersInExhibitor) findMesosMasters() (nodes []dcos.Node, err error) {
-	if f.getFn == nil {
-		return nodes, errors.New("could not initialize HTTP GET function. Make sure you set getFn in the constructor")
-	}
-	timeout := time.Second * 11
-	body, statusCode, err := f.getFn(f.url, timeout)
-	if err != nil {
-		return nodes, err
-	}
-	if statusCode != http.StatusOK {
-		return nodes, fmt.Errorf("GET %s failed, status code: %d", f.url, statusCode)
-	}
-
-	var exhibitorNodesResponse []exhibitorNodeResponse
-	if err := json.Unmarshal(body, &exhibitorNodesResponse); err != nil {
-		return nodes, err
-	}
-	if len(exhibitorNodesResponse) == 0 {
-		return nodes, errors.New("master nodes not found in exhibitor")
-	}
-
-	for _, exhibitorNodeResponse := range exhibitorNodesResponse {
-		nodes = append(nodes, dcos.Node{
-			Role:   dcos.MasterRole,
-			IP:     exhibitorNodeResponse.Hostname,
-			Leader: exhibitorNodeResponse.IsLeader,
-		})
-	}
-	return nodes, nil
-}
-
-func (f *findMastersInExhibitor) find() (nodes []dcos.Node, err error) {
-	nodes, err = f.findMesosMasters()
-	if err == nil {
-		logrus.Debug("Found masters in exhibitor")
-		return nodes, nil
-	}
-	// try next provider if it is available
-	if f.next != nil {
-		logrus.Warning(err)
-		return f.next.find()
-	}
-	return nodes, err
-}
-
-// NodesNotFoundError is a custom error called when nodes are not found.
-type NodesNotFoundError struct {
-	msg string
-}
-
-func (n NodesNotFoundError) Error() string {
-	return n.msg
-}
-
-// find agents in history service
-type findAgentsInHistoryService struct {
-	pastTime string
-	next     nodeFinder
-}
-
-func (f *findAgentsInHistoryService) getMesosAgents() (nodes []dcos.Node, err error) {
-	basePath := "/var/lib/dcos/dcos-history" + f.pastTime
-	files, err := ioutil.ReadDir(basePath)
-	if err != nil {
-		return nodes, err
-	}
-	nodeCount := make(map[string]int)
-	for _, historyFile := range files {
-		filePath := filepath.Join(basePath, historyFile.Name())
-		agents, err := ioutil.ReadFile(filePath)
-		if err != nil {
-			logrus.Errorf("Could not read %s: %s", filePath, err)
-			continue
-		}
-
-		unquotedAgents, err := strconv.Unquote(string(agents))
-		if err != nil {
-			logrus.Errorf("Could not unquote agents string %s: %s", string(agents), err)
-			continue
-		}
-
-		var sr agentsResponse
-		if err := json.Unmarshal([]byte(unquotedAgents), &sr); err != nil {
-			logrus.Errorf("Could not unmarshal unquotedAgents %s: %s", unquotedAgents, err)
-			continue
-		}
-
-		for _, agent := range sr.Agents {
-			if _, ok := nodeCount[agent.Hostname]; ok {
-				nodeCount[agent.Hostname]++
-			} else {
-				nodeCount[agent.Hostname] = 1
-			}
-		}
-
-	}
-	if len(nodeCount) == 0 {
-		return nodes, NodesNotFoundError{
-			msg: fmt.Sprintf("Agent nodes were not found in history service for the past %s", f.pastTime),
-		}
-	}
-
-	for ip := range nodeCount {
-		nodes = append(nodes, dcos.Node{
-			Role: dcos.AgentRole,
-			IP:   ip,
-		})
-	}
-	return nodes, nil
-}
-
-func (f *findAgentsInHistoryService) find() (nodes []dcos.Node, err error) {
-	nodes, err = f.getMesosAgents()
-	if err == nil {
-		logrus.Debugf("Found agents in the history service for past %s", f.pastTime)
-		return nodes, nil
-	}
-	// try next provider if it is available
-	if f.next != nil {
-		logrus.Warning(err)
-		return f.next.find()
-	}
-	return nodes, err
-}
-
-// find agents by resolving dns entry
-type findNodesInDNS struct {
-	forceTLS  bool
-	dnsRecord string
-	role      string
-	next      nodeFinder
-
-	// getFn takes url and timeout and returns a read body, HTTP status code and error.
-	getFn func(string, time.Duration) ([]byte, int, error)
-}
-
-func (f *findNodesInDNS) resolveDomain() (ips []string, err error) {
-	return net.LookupHost(f.dnsRecord)
-}
-
-func (f *findNodesInDNS) getMesosMasters() (nodes []dcos.Node, err error) {
-	ips, err := f.resolveDomain()
-	if err != nil {
-		return nodes, err
-	}
-	if len(ips) == 0 {
-		return nodes, errors.New("Could not resolve " + f.dnsRecord)
-	}
-
-	for _, ip := range ips {
-		nodes = append(nodes, dcos.Node{
-			Role: dcos.MasterRole,
-			IP:   ip,
-		})
-	}
-	return nodes, nil
-}
-
-func (f *findNodesInDNS) getMesosAgents() (nodes []dcos.Node, err error) {
-	if f.getFn == nil {
-		return nodes, errors.New("Could not initialize HTTP GET function. Make sure you set getFn in constractor")
-	}
-	leaderIps, err := f.resolveDomain()
-	if err != nil {
-		return nodes, err
-	}
-	if len(leaderIps) == 0 {
-		return nodes, errors.New("Could not resolve " + f.dnsRecord)
-	}
-
-	url, err := useTLSScheme(fmt.Sprintf("http://%s:5050/slaves", leaderIps[0]), f.forceTLS)
-	if err != nil {
-		return nodes, err
-	}
-
-	timeout := time.Second
-	body, statusCode, err := f.getFn(url, timeout)
-	if err != nil {
-		return nodes, err
-	}
-	if statusCode != http.StatusOK {
-		return nodes, fmt.Errorf("GET %s failed, status code %d", url, statusCode)
-	}
-
-	var sr agentsResponse
-	if err := json.Unmarshal(body, &sr); err != nil {
-		return nodes, err
-	}
-
-	for _, agent := range sr.Agents {
-		role := dcos.AgentRole
-
-		// if a node has "attributes": {"public_ip": "true"} we consider it to be a public agent
-		if agent.Attributes.PublicIP == "true" {
-			role = dcos.AgentPublicRole
-		}
-		nodes = append(nodes, dcos.Node{
-			Role: role,
-			IP:   agent.Hostname,
-		})
-	}
-	return nodes, nil
-}
-
-func (f *findNodesInDNS) dispatchGetNodesByRole() (nodes []dcos.Node, err error) {
-	if f.role == dcos.MasterRole {
-		return f.getMesosMasters()
-	}
-	if f.role != dcos.AgentRole {
-		return nodes, fmt.Errorf("%s role is incorrect, must be %s or %s", f.role, dcos.MasterRole, dcos.AgentRole)
-	}
-	return f.getMesosAgents()
-}
-
-func (f *findNodesInDNS) find() (nodes []dcos.Node, err error) {
-	nodes, err = f.dispatchGetNodesByRole()
-	if err == nil {
-		logrus.Debugf("Found %s nodes by resolving %s", f.role, f.dnsRecord)
-		return nodes, err
-	}
-	if f.next != nil {
-		logrus.Warning(err)
-		return f.next.find()
-	}
-	return nodes, err
-}
 
 // StartPullWithInterval will start to pull a DC/OS cluster health status
 func StartPullWithInterval(dt *Dt) {
@@ -364,9 +119,9 @@ func pullHostStatus(host dcos.Node, respChan chan<- *httpResponse, dt *Dt, wg *s
 	baseURL := fmt.Sprintf("http://%s:%d%s", host.IP, port, baseRoute)
 
 	// UnitsRoute available in router.go
-	url, err := useTLSScheme(baseURL, dt.Cfg.FlagForceTLS)
+	url, err := util.UseTLSScheme(baseURL, dt.Cfg.FlagForceTLS)
 	if err != nil {
-		logrus.Errorf("Could not read useTLSScheme: %s", err)
+		logrus.Errorf("Could not read UseTLSScheme: %s", err)
 		response.Status = http.StatusServiceUnavailable
 		host.Health = 3
 		response.Node = host
