+++ dcos/tools.go
@@ -1,4 +1,4 @@
-package api
+package dcos
 
 import (
 	"context"
@@ -6,20 +6,12 @@ import (
 	"io"
 	"io/ioutil"
 	"net/http"
-	netUrl "net/url"
 	"os"
 	"time"
 
-	"github.com/dcos/dcos-diagnostics/dcos"
-
 	"github.com/sirupsen/logrus"
-)
 
-const (
-	// _SYSTEMD_UNIT and UNIT are custom fields used by systemd to mark logs by the systemd unit itself and
-	// also by other related components. When dcos-diagnostics reads log entries it needs to filter both entries.
-	systemdUnitProperty = "_SYSTEMD_UNIT"
-	unitProperty        = "UNIT"
+	"github.com/dcos/dcos-diagnostics/util"
 )
 
 // GetHostname return a localhost hostname.
@@ -54,37 +46,15 @@ func (st *DCOSTools) GetNodeRole() (string, error) {
 	return st.Role, nil
 }
 
-func useTLSScheme(url string, use bool) (string, error) {
-	if use {
-		urlObject, err := netUrl.Parse(url)
-		if err != nil {
-			return "", err
-		}
-		urlObject.Scheme = "https"
-		return urlObject.String(), nil
-	}
-	return url, nil
-}
-
 // GetMesosNodeID return a mesos node id.
 func (st *DCOSTools) GetMesosNodeID() (string, error) {
 	// TODO(janisz): We need to decide if we need a context
 	return st.NodeInfo.MesosID(context.TODO())
 }
 
-// Help functions
-func isInList(item string, l []string) bool {
-	for _, listItem := range l {
-		if item == listItem {
-			return true
-		}
-	}
-	return false
-}
-
 func (st *DCOSTools) doRequest(method, url string, timeout time.Duration, body io.Reader) (responseBody []byte, httpResponseCode int, err error) {
 	if url != st.ExhibitorURL {
-		url, err = useTLSScheme(url, st.ForceTLS)
+		url, err = util.UseTLSScheme(url, st.ForceTLS)
 		if err != nil {
 			return responseBody, http.StatusBadRequest, err
 		}
@@ -96,7 +66,7 @@ func (st *DCOSTools) doRequest(method, url string, timeout time.Duration, body i
 		return responseBody, http.StatusBadRequest, err
 	}
 
-	client := NewHTTPClient(timeout, st.Transport)
+	client := util.NewHTTPClient(timeout, st.Transport)
 	resp, err := client.Do(request)
 	if err != nil {
 		return responseBody, http.StatusBadRequest, err
@@ -123,67 +93,34 @@ func (st *DCOSTools) GetTimestamp() time.Time {
 }
 
 // GetMasterNodes finds DC/OS masters.
-func (st *DCOSTools) GetMasterNodes() (nodesResponse []dcos.Node, err error) {
+func (st *DCOSTools) GetMasterNodes() (nodesResponse []Node, err error) {
 	finder := &findMastersInExhibitor{
 		url:   st.ExhibitorURL,
 		getFn: st.Get,
 		next: &findNodesInDNS{
 			forceTLS:  st.ForceTLS,
 			dnsRecord: "master.mesos",
-			role:      dcos.MasterRole,
+			role:      MasterRole,
 			next:      nil,
 		},
 	}
-	return finder.find()
+	return finder.Find()
 }
 
 // GetAgentNodes finds DC/OS agents.
-func (st *DCOSTools) GetAgentNodes() (nodes []dcos.Node, err error) {
+func (st *DCOSTools) GetAgentNodes() (nodes []Node, err error) {
 	finder := &findNodesInDNS{
 		forceTLS:  st.ForceTLS,
 		dnsRecord: "leader.mesos",
-		role:      dcos.AgentRole,
+		role:      AgentRole,
 		getFn:     st.Get,
-		next: &findAgentsInHistoryService{
-			pastTime: "/minute/",
-			next: &findAgentsInHistoryService{
-				pastTime: "/hour/",
+		next: &FindAgentsInHistoryService{
+			PastTime: "/minute/",
+			next: &FindAgentsInHistoryService{
+				PastTime: "/hour/",
 				next:     nil,
 			},
 		},
 	}
-	return finder.find()
-}
-
-// NewHTTPClient creates a new instance of http.Client
-func NewHTTPClient(timeout time.Duration, transport http.RoundTripper) *http.Client {
-	client := &http.Client{
-		Timeout: timeout,
-	}
-
-	if transport != nil {
-		client.Transport = transport
-	}
-
-	// go http client does not copy the headers when it follows the redirect.
-	// https://github.com/golang/go/issues/4800
-	client.CheckRedirect = func(req *http.Request, via []*http.Request) error {
-		for attr, val := range via[0].Header {
-			if _, ok := req.Header[attr]; !ok {
-				req.Header[attr] = val
-			}
-		}
-		return nil
-	}
-
-	return client
-}
-
-// open a file for reading, a caller if responsible to close a file descriptor.
-func readFile(fileLocation string) (r io.ReadCloser, err error) {
-	file, err := os.Open(fileLocation)
-	if err != nil {
-		return r, err
-	}
-	return file, nil
+	return finder.Find()
 }
